OPDS-PS v1.2 Support + a few bugfixes (#1869)
* Fixed up a localization lookup test case

* Refactored some webp to a unified method

* Cleaned up some code

* Expanded webp conversion for covers to all entities

* Code cleanup

* Prompt the user when they are about to delete multiple series via bulk actions

* Aligned Kavita to OPDS-PS 1.2.

* Fixed a bug where clearing metadata filter of series name didn't clear the actual field.

* Added some documentation

* Refactored how covert covers to webp works. Now we will handle all custom covers for all entities. Volumes and Series will not be touched but instead be updated via a RefreshCovers call. This will fix up the references much faster.

* Fixed up the OPDS-PS 1.2 attributes to only show on PS links

+++ UI/Web/src/app/_services/action.service.ts
@@ -537,12 +537,18 @@ export class ActionService implements OnDestroy {
    * @param chapters? Chapters, should have id
    * @param callback Optional callback to perform actions after API completes 
    */
-   deleteMultipleSeries(seriesIds: Array<Series>, callback?: VoidActionCallback) {
+   async deleteMultipleSeries(seriesIds: Array<Series>, callback?: BooleanActionCallback) {
+    if (!await this.confirmService.confirm('Are you sure you want to delete ' + seriesIds.length + ' series? It will not modify files on disk.')) {
+      if (callback) {
+        callback(false);
+      }
+      return;
+    }
     this.seriesService.deleteMultipleSeries(seriesIds.map(s => s.id)).pipe(take(1)).subscribe(() => {
       this.toastr.success('Series deleted');
 
       if (callback) {
-        callback();
+        callback(true);
       }
     });
   }
